+++ src/options/blacklist/actions.js
@@ -1,0 +1,7 @@
+import { createAction } from 'redux-act'
+
+export const setSiteInputValue = createAction('settings/setSiteInputValue')
+export const resetSiteInputValue = createAction('settings/resetSiteInputValue')
+export const setBlacklist = createAction('settings/setBlacklist')
+export const addSiteToBlacklist = createAction('settings/addSiteToBlacklist')
+export const removeSiteFromBlacklist = createAction('settings/removeSiteFromBlacklist')
